requirements.txt
alembic==1.13.1
aniso8601==9.0.1
apispec==6.3.1
apscheduler==3.10.4
astroid==3.0.2
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cached-property==1.5.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coverage[toml]==7.4.0
dill==0.3.7
dparse==0.6.3
et-xmlfile==1.1.0
flake8==6.1.0
flask==3.0.0
flask-apispec==0.11.4
flask-apscheduler==1.13.1
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-login==0.6.3
flask-marshmallow==0.15.0 
flask-migrate==4.0.5
flask-restful==0.3.10
flask-sqlalchemy==3.1.1
flask-testing==0.8.1
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
mako==1.3.0
markupsafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.29.0
mccabe==0.7.0
openpyxl==3.1.2
packaging==21.3
platformdirs==4.1.0
pluggy==1.3.0
psycopg2==2.9.9
pycodestyle==2.11.1
pyjwt==2.8.0
pylint==3.0.3
pyparsing==3.1.1
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyyaml==6.0.1
referencing==0.32.0
requests==2.31.0
responses==0.24.1
rpds-py==0.16.2
ruamel-yaml==0.18.5
ruamel-yaml-clib==0.2.8
safety==2.3.5
six==1.16.0
sqlalchemy==2.0.25
sqlalchemy-history==2.1.4
sqlalchemy-utils==0.41.1
tomlkit==0.12.3
typing-extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
webargs==8.3.0
werkzeug==3.0.1
wheel==0.42.0